IPv6: printing a socket: it is not an int
When printing a socket under windows, we get a warning:
a SOCKET is a "long long unsigned" (under Windows)

Cast it into a an "int" when printing.
On most platform it is already an "int", under Windows we don't print
the whole SOCKET.
Using a "long" does not help, since a "long" under Windows is 32 bits,
the same as an "int".

+++ modules/ca/src/client/tcpiiu.cpp
@@ -1059,7 +1059,7 @@ void tcpiiu::show ( unsigned level ) const
             this->_receiveThreadIsBusy );
     }
     if ( level > 2u ) {
-        ::printf ( "\tvirtual circuit socket identifier %d\n", this->sock );
+        ::printf ( "\tvirtual circuit socket identifier %d\n", (int)this->sock );
         ::printf ( "\tsend thread flush signal:\n" );
         this->sendThreadFlushEvent.show ( level-2u );
         ::printf ( "\tsend thread:\n" );
